src/requirements.txt
Django==4.2.11
django-extensions==3.2.3
whitenoise==6.6.0
python-dotenv==0.21.1
psycopg2-binary
urllib3==1.26.15
requests==2.31.0
SpeechRecognition==3.10.1
pydub==0.25.1
channels==4.0.0
daphne==4.1.0
channels_redis==4.2.0
uvicorn[standard]
websocket-client
watchfiles
stripe
django-widget-tweaks